Add nimble-select Angular guidance for compareWith function (#2699)
# Pull Request

## 🤨 Rationale

I recently investigated a
[bug](https://ni.visualstudio.com/DevCentral/_workitems/edit/3244716)
filed for the `nimble-select` about incorrect behavior when using
`filter-mode` of `manual`:

> When using `nimble-select` in an Angular reactive form with
`filter-mode="manual"`, the component's visual display value is cleared
if the list of `nimble-list-option` elements is dynamically replaced.
This occurs even if the new list contains an option that corresponds to
the selected value and the underlying `formControl` value remains
correct.
> 
> **Steps to Reproduce**
> Create an Angular component with a formControl.
> Bind this formControl to a <nimble-select> with filter-mode="manual".
> Populate the select with <nimble-list-option> elements using *ngFor
and bind objects to the options using [ngValue].
> Implement a filter function (e.g., on the (filter-input) event) that
creates a new array of options.
> Run the application and select an option from the dropdown. The select
correctly displays the name of the selected object.
> Type into the filter. The list of options is rebuilt.
> Expected Behavior
> If the currently selected object is present in the new, filtered list
of options, the nimble-select should continue to display its name/label.
>  **Actual Behavior**
> The nimble-select's display value is cleared and it reverts to showing
the placeholder, even though formControl.value still holds the correct
selected object.

Based on my investigation, the "binding to options with [ngValue]" was
the important part. I could reproduce the behavior described
(`nimble-select` reverting to showing the placeholder after updating the
options). However specifying a `compareWith` function on the
`nimble-select` was sufficient to fix the problem.

## 👩‍💻 Implementation

Update `nimble-select` Storybook docs with guidance about specifying a
`compareWith` function if binding to options with `ngValue`.

## 🧪 Testing

PR build.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.