Add jasmine-extensions and re-enable intermittent tests (#2701)
# Pull Request

## 🤨 Rationale

Nimble has a lot of intermittent tests that were disabled.

## 👩‍💻 Implementation

Implemented a script that enables Jasmine tests to do retries. It's a
little blunt and just re-runs `it` blocks a number of times but notably
does not call / recall associated `beforeEach` and `afterEach`
statements. We haven't generally written tests with the assumption that
`beforeEach` is called once while `it` statements may be called one or
more times so this is a bit of a shot in the dark that if a second run
passes then it's good enough 🤷‍♂️.

## 🧪 Testing

- Added a couple tests to jasmine-extensions to illustrate behavior
- Previous intermittent tests will be re-enabled to see if retries help
😬 (many of these are likely stable now just because we have a newer
version of webkit but we haven't run #1747 in a while to check)
  - resolves https://github.com/ni/nimble/issues/2169
  - resolves https://github.com/ni/nimble/issues/2603
  - resolves https://github.com/ni/nimble/issues/2274
  - resolves https://github.com/ni/nimble/issues/2269
  - resolves https://github.com/ni/nimble/issues/2426
  - resolves https://github.com/ni/nimble/issues/2272
  - resolves https://github.com/ni/nimble/issues/2170

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.