Exercise intermittent tests (#2703)
# Pull Request

## 🤨 Rationale

Recently added test retry code on the CI to reduce intermittency in
#2701
Have the CI run on main every hour to exercise for intermittency on the
CI.
We can let it run for a while, maybe a week or so?

## 👩‍💻 Implementation

Add a [cron schedule
trigger](https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#schedule)
to run the CI every hour.

## 🧪 Testing

The testing is the CI

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.