Expose "disabled" attribute in Angular directive for the nimble-number-field (#2733)
# Pull Request

## 🤨 Rationale

It was pointed out that you cannot bind to the "disabled" property on
the nimble-number-field in Angular. This is because the directive is
missing the "disabled" input.

## 👩‍💻 Implementation

Add "disabled" input on the Angular number-field directive following
existing patterns.

## 🧪 Testing

- Added unit tests following existing patterns

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.