Checkbox directive fix (#2737)
# Pull Request

## 🤨 Rationale

Fixes #2734

## 👩‍💻 Implementation

Resetting the form control sets the value to `null` (unless
`nonNullable` is configured on the form control). However, the `change`
event fired from the checkbox due to the programmatic value change
causes the value to be set again to `false`. Therefore, to solve the
problem, I added code to ignore change events fired from the checkbox
while the form was programmatically setting the value.

Ultimately, it would be nice to solve the root cause of the problem
which is that the checkbox fires a `change` event when the value is
programmatically set. However, this is a larger problem to tackle -- not
only for the checkbox but across all/most FAST-based components.
Therefore, I think this is an acceptable solution for now, and it is
aligned with a change that was put into the table to resolve a similar
problem related to row selection (see
https://github.com/ni/nimble/blob/main/packages/nimble-components/src/table/components/group-row/index.ts#L75-L79)

## 🧪 Testing

- Manually tested in an Angular app
- Added new unit tests for the checkbox CVA that creates a form control
(previous checkbox CVA tests only exercised `ngModel` binding code
paths)

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>