[HLD] Table Selection Behavior with Action Button  (#2732)
# Pull Request

## 🤨 Rationale

For some tables, there is a need to open the action menu without
changing the selection. This HLD proposes a way to support multiple
selection modes in a table when opening an action menu. This is the
design for #2708 .

## 👩‍💻 Implementation

N/A

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>