Fix bug where dialogs and drawers can close via ESC when preventDismiss is true (#2759)
# Pull Request

## 🤨 Rationale

Resolves #2705 

With this change, the `nimble-dialog` and `nimble-drawer` no longer
close when "ESC" is pressed twice when `preventDismiss` is set to
`true`. They both still close on the first "ESC" press when
`preventDismiss` is `false`.

## 👩‍💻 Implementation

In both the dialog and drawer, configure `closedby` on the inner
`dialog` element to be `"none"` when `preventDismiss` is `true` and to
be `"closerequest"` when `preventDismiss` is `"none"`.

## 🧪 Testing

- Added unit tests for the dialog and drawer to ensure `closedby` has
the correct value
- Manually tested via storybook

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.