Rename DisabledReadOnlyStates to ManipulationStates (#2762)
# Pull Request

## 🤨 Rationale

The name `DisabledReadOnlyState` was challenging to follow during a
recent update. Came up with a name that avoids the words "disabled" and
"readOnly".

## 👩‍💻 Implementation

Using the term `ManipulationStates` to represent `disabled` / `readOnly`
/ `appearanceReadOnly`. The term `ManipulabilityStates` may be be
slightly better but I found `Manipulation` easier to scan and type.

## 🧪 Testing

Rely on CI to report no diffs.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.