Add work around to table to fix action-menu bug in Firefox (#2760)
# Pull Request

## 🤨 Rationale

Provides a work around for the action menu issue reported in #2744

## 👩‍💻 Implementation

This doesn't solve the root cause of the issue, but it does avoid it (in
a way that I don't fully understand).

The change in this PR sets a row's `currentActionMenuColumn` to
`undefined` when the action menu closes, which removes the `slot`
element for the menu within row.

## 🧪 Testing

Manually verified that the action menu can be opened on multiple rows in
the table without an exception being thrown and without data vanishing
from the table.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>