Use Sass @use instead of deprecated @import (#2763)
# Pull Request

## 🤨 Rationale

The Sass `@import` statement is
[deprecated](https://sass-lang.com/blog/import-is-deprecated/) in favor
of `@use` and `@forward`. In recent versions of Dart Sass (which will be
used when upgrading to Angular 19), use of `@import` leads to noisy
build warnings.

## 👩‍💻 Implementation

We could just silence the build warnings via a build option, but it
seems preferable to just do the replacement. The `@use` statement
provides support for namespacing imported variables, but to avoid having
to modify every use of a variable, we're opting to import the variables
in the global namespace per SCSS file, i.e. `@use "<tokens_file>" as *`.
See [official
documentation](https://sass-lang.com/documentation/at-rules/use/#choosing-a-namespace)
for details on how namespacing with `@use` works.

## 🧪 Testing

SystemLinkShared libs build.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>