Fix link to Example Client App in README (#2775)
# Pull Request

## 🤨 Rationale

Fixes #2776 - A recent Renovate update added a directory field to the
package.json for most components (a few already had it specified).

While NPM
[suggests](https://docs.npmjs.com/cli/v11/configuring-npm/package-json)
that monorepos should use this field "if the package.json for your
package is not in the root directory (for example if it is part of a
monorepo), you can specify the directory in which it lives…", it breaks
our absolute relative paths strategy.

## 👩‍💻 Implementation

Removed directory path from all package.json files

## 🧪 Testing

- [x] Confirmed that existing links continue to work in Github. 
- [ ] Confirm that NPM readme links are correct


## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.