Eslint 9 Format Result (#2774)
# Pull Request

## 🤨 Rationale

This PR is the result of running format and cleaning up the result based
on the changes in #2773

Note: This PR is intended to review the source changes that are made and
will be merged into the `eslint-v9` branch, not `main`.

## 👩‍💻 Implementation

Ran `npm run format` and tweaked the results a but to remove whitespace
and address a couple rules that did not have automatic formatters.

## 🧪 Testing

Rely on CI to report no visual diffs and pass.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: rajsite <1588923+rajsite@users.noreply.github.com>