Explicitly mark Angular components/directives/pipes as non-standalone (#2785)
# Pull Request

## 🤨 Rationale

Separate this noisy change from the Angular 19 upgrade changes.
(Components are standalone by default in Angular 19, so these changes to
the decorators are no-ops in v18, but necessary in v19.)

## 👩‍💻 Implementation

This was generated by doing `ng update` (automatic migration), then I
separated the changes to submit.

## 🧪 Testing

Builds and tests pass.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>