Always format doubles in razor files using the invariant culture (#2778)
# Pull Request

## 🤨 Rationale

Fixes #2743

## 👩‍💻 Implementation

When you assign a `double` value to an attribute in a razor file (e.g.
`max="@Max"`), it formats the value by calling `double.ToString()`. By
default, `ToString()` uses the current culture. This means a
non-integral value can have a comma as the decimal separator rather than
a dot, based on the current culture. This isn't valid for HTML. All
numbers have to use a dot for a decimal separator.

I searched all the razor.cs files for `double` properties, then looked
for their use in the razor files. I've updated all of them to explicitly
format the values with the invariant culture:

- `NumberField` - overriding `FormatValueAsString` (from
`NimbleInputBase`) to use invariant culture and calling that from the
razor file. Also changed `@CurrentValue` to `@CurrentValueAsString`.
- `NimbleTableColumn*` - introduced a new base class,
`NimbleFractionalTableColumn`, that provides the `FractionalWidth` and
`MinPixelWidth` properties, as well as two new readonly properties,
`FractionalWidthAsString` and `MinPixelWidthAsString`. Using the new
properties in the razor file instead.
- `NimbleWaferMap` - explicitly formatting `GridMin/MaxX/Y` and
`MaxCharacters` (no idea why this is a `double`) with
`ToString(CultureInfo.InvariantCulture)` in the razor file.

## 🧪 Testing

Set my browser's language to German, then tested in the Blazor example
app.

Added new unit test cases.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>