Upgrade storybook dependencies to latest patch version (#2798)
# Pull Request

## 🤨 Rationale

Chromatic sent an email indicating there was a security vulnerability
with the version of storybook we're using. From a surface inspection I
don't think it affects us but it seems prudent to update to a version
where it has been addressed.

## 👩‍💻 Implementation

Update all storybook packages to latest and run npm install.

## 🧪 Testing

Cursory inspection of local storybook build.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.