Enable dark mode for storybook UI (#2802)
# Pull Request

## 🤨 Rationale

For the Storybook UI the PR follows the configured system theme to
enable light or dark mode for the manager ui and the docs ui

## 👩‍💻 Implementation

- Add a dark theme to the storybook theme
- Use storybook's `getPreferredColorScheme` similar to how the storybook
`normal` theme is implemented
- Avoided changing the default selected theme for the canvas to the
system theme as that had funny behavior. It stays defaulted to light
theme.

## 🧪 Testing

Tried storybook locally and rely on no Chromatic diffs.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.